fixing ut after merge conflicts, and organized report structs
